fix: rename directory to inputs Signed-off-by: Philippe Martin <phmartin@redhat.com>
