+++ packages/ui/src/lib/inputs/SearchInput.svelte
